Create a function that replicates an analysis for all time series in a dataset

This wrapper takes as input, some analysis function, fun, along with additional optional arguments, and returns a function that will run that analysis function on all the time series in a dataset. The return object from that function combined all the relevant output, and also information on the call, for post-processing that is suitable by itself, or as part of a Drake plan.

analysis_wrapper(fun, ...)

Arguments

fun
the analysis function

...
extra params to be passed to the analysis function

Value

a function that takes in a single argument, dataset, and returns a tibble with these columns:

results

a combined results data.frame

metadata

the metadata component of the original dataset

dataset

the name of the dataset

method

the name of the analysis function

args

a list of optional args to

method

Examples

# NOT RUN {
  sgs_data <- MATSS::get_sgs_data()
  summarize_dataset <- analysis_wrapper(ts_summary)
  summarize_dataset(sgs_data)
# }