Collect Analyses Together into a Tibble

This is a helper function to accompany plan_analyses: it is necessary to collect all of the results that are produced by the drake plan.

This function strives to be intelligent about the format of the individual results. For output from analysis_wrapper that already has information about the method and dataset, we can just combine them. Otherwise, we parse the name of the object for the method and the dataset, to format into a structure similar to the output from analysis_wrapper.

collect_analyses(list_of_results)

Arguments

list_of_results
the list of objects

Value

a drake plan (i.e. a tibble) specifying the targets and commands for all the analyses and the collected results (grouping the outputs from each method into a single list)