Create a research compendium for MATSS

Create an R package, using use_package, then perform the following additional actions:

add various package dependencies, including "MATSS"

add template analysis and pipeline files

create_MATSS_compendium(
  path,
  fields = NULL,
  rstudio = rstudioapi::isAvailable(),
  open = interactive()
)

Arguments

path
A path. If it exists, it is used. If it does not exist, it is created, provided that the parent path exists.

fields
A named list of fields to add to DESCRIPTION, potentially overriding default values. See use_description() for how you can set personalized defaults using package options

rstudio
If TRUE, calls use_rstudio() to make the new package or project into an RStudio Project. If FALSE and a non-package project, a sentinel .here file is placed so that the directory can be recognized as a project by the here or rprojroot packages.

open
If TRUE, activates the new project:

If RStudio desktop, the package is opened in a new session.

If on RStudio server, the current RStudio project is activated.

Otherwise, the working directory and active project is changed.

Value

Path to the newly created package, invisibly.