get portal rodent data Import Portal rodent data using portalr functions. Currently returns rodent data formatted appropriately for LDA analysis.

get_portal_rodents(time_or_plots = "plots", treatment = "control",
  type = "Rodents")

Arguments

time_or_plots
select whether to: (1) "time" == get the data for the entire timespan of the experiment, or (2) "plots" == just the time period with consistent treatments

treatment
"control" or "exclosure" treatments

type
type of animals to get: "Rodents" or restrict to "Granivores"